Make validated tables into slabs.

+++ include/bitcoin/database/tables/indexes/strong_bk.hpp
@@ -55,7 +55,7 @@ struct strong_bk
             return code == other.code;
         }
 
-        coding code{};
+        coding::integer code{};
     };
 };
